Add support for HTTPX instrumentation
This builds off of the initial idea in [this issue][1] of using the
transport API in `httpx`. This allows for using [custom transports][2].

One issue with this current implementation is that there are few more
[changes coming][3] in 0.18.0 that this will probably want to use. So
maybe it will make sense to finalize this on that version before
releasing it. Want to get some eyes on this sooner.

Resolves #263

[1]: https://github.com/encode/httpx/issues/1264#issuecomment-732034750
[2]: https://www.python-httpx.org/advanced/#custom-transports
[3]: https://github.com/encode/httpx/pull/1522